backend/generated_agents/6bc654b0-235c-4d96-a8b1-b25eb7a9b006/requirements.txt
crewai
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
